removed ViewModelServices
