+++ src/Rutracker.Server.DataAccessLayer/Entities/UserRoles.cs
@@ -2,10 +2,7 @@
 {
     public static class UserRoles
     {
-        public static class Names
-        {
-            public const string User = "user";
-            public const string Admin = "admin";
-        }
+        public const string User = "user";
+        public const string Admin = "admin";
     }
 }
