+++ src/Rutracker.Server.WebApi/Controllers/TorrentsController.cs
@@ -1,8 +1,10 @@
 ﻿using System;
+using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
+using AutoMapper;
+using Rutracker.Server.BusinessLayer.Interfaces;
 using Rutracker.Server.WebApi.Controllers.Base;
-using Rutracker.Server.WebApi.Interfaces;
 using Rutracker.Shared.Models.ViewModels.Shared;
 using Rutracker.Shared.Models.ViewModels.Torrent;
 
@@ -15,11 +17,13 @@ namespace Rutracker.Server.WebApi.Controllers
     /// <response code="404">If the item is null.</response>
     public class TorrentsController : BaseApiController
     {
-        private readonly ITorrentViewModelService _torrentViewModelService;
+        private readonly ITorrentService _torrentService;
+        private readonly IMapper _mapper;
 
-        public TorrentsController(ITorrentViewModelService torrentViewModelService)
+        public TorrentsController(ITorrentService torrentService, IMapper mapper)
         {
-            _torrentViewModelService = torrentViewModelService ?? throw new ArgumentNullException(nameof(torrentViewModelService));
+            _torrentService = torrentService ?? throw new ArgumentNullException(nameof(torrentService));
+            _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
         /// <summary>
@@ -31,7 +35,16 @@ namespace Rutracker.Server.WebApi.Controllers
         [HttpPost(nameof(Pagination))]
         public async Task<PaginationResult<TorrentViewModel>> Pagination(int page, int pageSize, FilterViewModel filter)
         {
-            return await _torrentViewModelService.TorrentsAsync(page, pageSize, filter);
+            var torrents = await _torrentService.ListAsync(page, pageSize, filter.Search, filter.SelectedForumIds, filter.SizeFrom, filter.SizeTo);
+            var count = await _torrentService.CountAsync(filter.Search, filter.SelectedForumIds, filter.SizeFrom, filter.SizeTo);
+
+            return new PaginationResult<TorrentViewModel>
+            {
+                Page = page,
+                PageSize = pageSize,
+                TotalPages = (int)Math.Ceiling(count / (double)pageSize),
+                Items = _mapper.Map<TorrentViewModel[]>(torrents)
+            };
         }
 
         /// <summary>
@@ -41,7 +54,9 @@ namespace Rutracker.Server.WebApi.Controllers
         [HttpGet]
         public async Task<TorrentDetailsViewModel> Get(long id)
         {
-            return await _torrentViewModelService.TorrentAsync(id);
+            var torrent = await _torrentService.FindAsync(id);
+
+            return _mapper.Map<TorrentDetailsViewModel>(torrent);
         }
 
         /// <summary>
@@ -51,7 +66,17 @@ namespace Rutracker.Server.WebApi.Controllers
         [HttpGet(nameof(Titles))]
         public async Task<FacetResult<string>> Titles(int count)
         {
-            return await _torrentViewModelService.ForumFacetAsync(count);
+            var facets = await _torrentService.ForumsAsync(count);
+
+            return new FacetResult<string>
+            {
+                Items = facets.Select(x => new FacetViewModel<string>
+                {
+                    Id = x.Item1.ToString(),
+                    Value = x.Item2,
+                    Count = x.Item3
+                }).ToArray()
+            };
         }
     }
 }
